Add pallet to own crate
