+++ integration-tests/public/runtime-call-contract/e2e_tests/pallet-contract-caller/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "pallet-contract-caller"
+version = "0.1.0"
+description = "Demonstrate calling an ink! contract from a pallet"
+authors = ["Substrate DevHub <https://github.com/substrate-developer-hub>"]
+homepage = "https://substrate.io"
+edition.workspace = true
+license = "MIT-0"
+publish = false
+repository = "https://github.com/paritytech/ink"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+codec = { package = "parity-scale-codec", workspace = true, default-features = false, features = [
+    "derive",
+] }
+scale-info = { workspace = true, default-features = false, features = ["derive"] }
+frame-support = { workspace = true, default-features = false }
+frame-system = { workspace = true, default-features = false }
+
+pallet-contracts = { workspace = true, default-features = false }
+flipper-traits = { workspace = true, default-features = false }
+ink = { workspace = true, features = ["no-panic-handler", "no-allocator"] }
+
+[features]
+default = ["std"]
+std = [
+    "codec/std",
+    "frame-support/std",
+    "frame-system/std",
+    "scale-info/std",
+    "pallet-contracts/std",
+    "ink/std"
+]
